Split kdtree decl and impl.
* Depend on all lib/*.o files in binaries.
* Move split box implementation to types.
* Ignore radiosity binary.
