+++ tests/test_kdtree.cpp
@@ -7,24 +7,6 @@
 #include <iostream>
 
 
-TEST_CASE("Split box at plane", "[box]") {
-    Box box{{0, 0, 0}, {2, 2, 2}};
-
-    Box l, r;
-    std::tie(l, r) = split(box, Axis::X, 1);
-    REQUIRE(l == (Box{{0, 0, 0}, {1, 2, 2}}));
-    REQUIRE(r == (Box{{1, 0, 0}, {2, 2, 2}}));
-
-    std::tie(l, r) = split(box, Axis::Y, 1);
-    REQUIRE(l == (Box{{0, 0, 0}, {2, 1, 2}}));
-    REQUIRE(r == (Box{{0, 1, 0}, {2, 2, 2}}));
-
-    std::tie(l, r) = split(box, Axis::Z, 1);
-    REQUIRE(l == (Box{{0, 0, 0}, {2, 2, 1}}));
-    REQUIRE(r == (Box{{0, 0, 1}, {2, 2, 2}}));
-}
-
-
 TEST_CASE("Trivial smoke test", "[kdtree]")
 {
     auto tri = random_triangle();
@@ -142,10 +124,16 @@ TEST_CASE("KDTree stress test", "[kdtree]")
             }
         }
     }
+
     std::cerr << "Runtime  : " << runtime_ms << "ms" << std::endl;
     std::cerr << "# Hits   : " << hits_fast << std::endl;
     std::cerr << "Rays/sec : " << 1000. * RAYS_COUNT / runtime_ms << std::endl;
     std::cerr << std::endl;
+
+    // empirical values
+    REQUIRE(tree.height() == 10);
+    REQUIRE(tree.size() == 115);
+    REQUIRE(hits_fast == 4);
 }
 
 TEST_CASE("Test cube in kdtree", "[kdtree]") {
