+++ tests/tests.mk
@@ -13,7 +13,7 @@ define GENERIC_TEST
 genfiles/$(1).o: tests/$(1).cpp lib/*.h
 	$(CXX) $(TESTS_CXXFLAGS) -c $$< -o $$@
 
-genfiles/$(1): genfiles/$(1).o genfiles/main_test.o lib/types.o lib/triangle.o
+genfiles/$(1): genfiles/$(1).o genfiles/main_test.o $(LIB_OBJS)
 	$(CXX) $(LDFLAGS) $$^ -o $$@
 
 .PHONY: test
